Reorder deployment stages and separate admin node
This task is in preparation for a shift to an ansible
playbook for preparing nodes with whatever required
packages and workarounds that are not covered by Kargo.

Admin node will now be configurable to be "local" or ADMIN_IP,
but still defaults to the first slave node if fuel_devops is used
for provisioning.

Change-Id: I5fc913315142280590b7e7cf5d56b2d1146b56c8